Update docs command

+++ scripts/docs/docs.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env sh
 
+if [[ `git status --porcelain` ]]; then
+  echo  "Working directory must be clean"
+  exit 1
+fi
 echo Removing current docs folder...
 rm -rf docs
 echo Building new docs...
@@ -12,3 +16,12 @@ touch docs/.nojekyll
 echo Adding CNAME file...
 echo jsonapy.pycolore.fr > docs/CNAME
 echo Done.
+echo Switching to gh-pages branch...
+git add docs
+git stash
+git checkout gh-pages
+git stash pop --quiet
+git checkout --theirs .
+git add docs
+git stash drop
+echo Ready to commit.
